requirements.txt
feedparser==6.0.11
python-docx==1.1.2
deep-translator==1.11.4
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
